chore(incidents): Rename Incident -> ServiceIncident
Since we now have Incidents as a product concept in Sentry we need to
rename service incidents on the Sentry service which were previously
referred to as Incidents as well.
